Create and display DAG based on OpenMRS role/privs
Able to read role/privilege data using OpenMRS API and use it to create
a directed acyclic graph (DAG) made up of cryptographic nodes and edges.
The endpoint "/openmrs/module/dynaswap/test.page" will display a page
where a logged in user can click a button to create and display the DAG.

Note: At this point, the DAG will not yet save anything to the database
or modify it.